moved post components into post folder
