+++ Tabloid/client/src/components/posts/PostList.js
@@ -1,5 +1,5 @@
 import React, { useContext, useEffect, useState, useRef } from "react";
-import { PostContext } from "../providers/PostProvider";
+import { PostContext } from "../../providers/PostProvider";
 import PostCard from "./PostCard";
 import { useHistory, Link } from "react-router-dom";
 
@@ -12,22 +12,21 @@ const PostList = () => {
         getAllPosts();
     }, []);
 
-    // console.log("hello!!!")
-    // console.log(JSON.parse(sessionStorage.getItem("userProfile")).id)
+
 
     return (
         <div className="container">
             <div className="row justify-content-center">
                 <div className="cards-column">
                     {posts.map((post) => (
-                        <>
+                        <div key={post.id}>
                             <Link to={`/posts/${post.id}`}>
                                 <strong>{post.title}</strong>
                             </Link>
                             <p>Author: {post.userProfile.firstName} {post.userProfile.lastName}</p>
                             <p>Category: {post.category.name}</p>
 
-                        </>
+                        </div>
                     ))}
                 </div>
             </div>
