+++ Tabloid/client/src/components/posts/UserPostList.js
@@ -1,5 +1,5 @@
 import React, { useContext, useEffect, useState, useRef } from "react";
-import { PostContext } from "../providers/PostProvider";
+import { PostContext } from "../../providers/PostProvider";
 import PostCard from "./PostCard";
 import { useHistory, Link } from "react-router-dom";
 
@@ -7,8 +7,7 @@ const UserPostList = () => {
     const { posts, getAllUserPosts } = useContext(PostContext);
     const history = useHistory();
 
-    console.log("hello!!!")
-    console.log(JSON.parse(sessionStorage.getItem("userProfile")).id)
+
 
     useEffect(() => {
         getAllUserPosts(JSON.parse(sessionStorage.getItem("userProfile")).id);
@@ -19,14 +18,14 @@ const UserPostList = () => {
             <div className="row justify-content-center">
                 <div className="cards-column">
                     {posts.map((post) => (
-                        <>
+                        <div key={post.id}>
                             <Link to={`/posts/${post.id}`}>
                                 <strong>{post.title}</strong>
                             </Link>
                             <p>Author: {post.userProfile.firstName} {post.userProfile.lastName}</p>
                             <p>Category: {post.category.name}</p>
 
-                        </>
+                        </div>
                     ))}
                 </div>
             </div>
